feat:调整项目结构，优化多个代码结构问题，当前流程全部通过测试

+++ GateServer/model/User.h
@@ -5,7 +5,7 @@
 #include <cstdint>
 
 struct UserInfo {
-    int uid = -1;
+    int64_t uid = -1;
     std::string username;
     std::string email;
     std::string passwordHash;
